EJS + Purchase Page + Formatting
EJS lets us render html files with simple paths and use parameters in html instead of js only. html files are now in views. static files that do not change are in public. Moved server.js to root. I changed the formatting to be like what was done in class. The purchase page uses a parameter to know which item you are buying. It is connected to the store page by clicking on an image or title of an item.